+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Library mods are now shown by default
 - Also print the extension name when it's missing and a mod requires it
 - For CCLoader3, the mod installation dir is now `modloader.config.modsDirs`
+- Improved behavior when there's no internet
 
 ### Fixed